requirements.txt
absl-py==0.14.1
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==4.2.4
cattrs==22.1.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
clang==5.0
click==8.1.3
colorama==0.4.5
cycler==0.10.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
discord==2.0.0
discord.py==2.0.1
entrypoints==0.4
exceptiongroup==1.0.0rc9
executing==0.9.1
fastf1==2.3.0
fastjsonschema==2.16.1
Flask==2.2.2
flatbuffers==1.12
frozenlist==1.3.1
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
heroku==0.1.4
idna==3.3
importlib-metadata==5.0.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.0.1
jsonschema==4.8.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mnist==0.2.2
multidict==6.0.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.3
oauthlib==3.1.1
opendatasets==0.1.22
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.18.1
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyparsing==2.4.7
pyrsistent==0.18.1
PySimpleGUI==4.48.0
PySimpleGUIWeb==0.39.0
python-dateutil==1.5
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.1
pywin32==304
pywinpty==2.0.6
pyzmq==23.2.0
remi==2020.3.10
requests==2.28.1
requests-cache==0.9.6
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
seaborn==0.12.0
Send2Trash==1.8.0
signalr-client-aio==0.0.1.6.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.10.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.15.0
text-unidecode==1.3
thefuzz==0.19.0
threadpoolctl==2.2.0
timple==0.1.5
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing-extensions==3.7.4.3
url-normalize==1.4.3
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
Werkzeug==2.2.2
wrapt==1.12.1
yarl==1.8.1
zipp==3.8.1
